// server/src/common/enums.ts
export enum RequestType {
	Corrective = "corrective",
	Preventive = "preventive"
}

export enum RequestState {
	New = "new",
	InProgress = "in_progress",
	Repaired = "repaired",
	Scrap = "scrap"
}

export enum EquipmentStatus {
	Active = "active",
	Scrapped = "scrapped"
}

export enum TeamRole {
	Manager = "manager",
	Technician = "technician"
}